requirements.txt
numpy==2.3.4
scrapy==2.11.1
lxml==5.3.0
beautifulsoup4==4.12.3
requests==2.32.3
pydantic==2.9.2
python-dotenv==1.0.1
kafka-python==2.0.2
fastapi==0.115.5
uvicorn==0.32.0
qdrant-client==1.12.0
openai==1.52.2
tenacity==9.0.0
pymongo==4.10.1
